Migrate things to the wiki
Update the README now that the documentation is on the wiki.
Also hyperlink to Web Polygraph, since it's rather important.

Rewrite INSTALL to be a pointer to the wiki now its content has moved
there and will be improved.
Remove doc folder now its only content has been moved to the wiki.