Add post{inst,rm} and init scripts
Tweak the init script to be slightly more conventional in its naming.

Install the init script as part of each component package.

Add a postinst and postrm to each component that creates/disables an
appropriate user.

Also fix incorrect use of --interactive in a few places in the rules.

Everything appears to be hooking up correctly, but not yet tested
installing it.