Fixes for packaging and init
Fix a stupid typo in the init script that was badly misconfiguring the
UNIX-y name of the component.

Make the Plumage base install directory default-configurable, and have
the Debian package make use of this. (I still want to avoid making the
init script too Debian-packaging-specific, since it should work with
other deployments to SysV-like systems.)

Fix a stupid syntax error in postinst (and correct a comment).