Start improving Plumage UI error reporting
Turn on backtraces, and provide a template to show exceptions to the
browser, rather than the default one.

Slightly improve the formatting of the usual API helper exceptions.

Now at least when you try to go to a configuration that doesn't exist,
you can see there was an underlying 404 behind the 500 without tailing
the logs, but there's a long way to go yet.