Do not provide anything in LSB init headers
Since the same init script is used for all roles (adaptive depending on
how it is called), remove the LSB 'Provides' header value, since it is
incorrect for multiple services on the same machine to provide the same
token, and modern versions of insserv (e.g. Debian Jessie) will actually
detect and fail on this.

Providing no token at all avoids needing either manual installations or
the Debian rules from having to modify this value per-script, and there
is little use in having plumage as a provided service since no other
service should be depending on it.